Create a new open mct Plan API. Move execution monitoring providers to the plan API.
